# buzzbar-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuzzBar Backend API
  version: "1.0.0"
  description: |
    Dev Phase 1 API (stable).

    Stability guarantee:
    - `/api/v1/**` is considered stable after Phase 1.
    - Future changes must be additive only. Breaking changes require `/api/v2`.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/v1/admin/kyc/queue:
    get:
      tags: [AdminKYC]
      summary: KYC review queue
      security: [{ AdminBearerAuth: [] }]
      responses:
        "200":
          description: OK
  /api/v1/admin/kyc/{userId}:
    get:
      tags: [AdminKYC]
      summary: View user KYC (signed image URLs)
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
  /api/v1/admin/kyc/{userId}/approve:
    post:
      tags: [AdminKYC]
      summary: Approve user KYC
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
  /api/v1/admin/kyc/{userId}/reject:
    post:
      tags: [AdminKYC]
      summary: Reject user KYC
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    UserBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...